# svehi-backend/src/main/resources/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SVEHI Modernización API
  description: API para el sistema de Gestión de Vehículos e Históricos (SVEHI)
  version: 1.0.0
  contact:
    name: Arquitectura SVEHI
servers:
  - url: http://localhost:8080/svehi
    description: Local Server
...
paths:
...
  /v1/auth/login-cert:
    post:
      tags:
        - Auth
      summary: Iniciar sesión con certificado @firma
      operationId: loginWithCertificate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginCertificateRequestVO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioVO'
        '401':
          description: Unauthorized (Usuario no encontrado o inactivo)
  /v1/usuarios:
    get:
      tags:
        - Usuarios
      summary: Listar usuarios
      operationId: getUsuarios
      parameters:
        - name: nif
          in: query
          description: Filtrar por NIF
          schema:
            type: string
        - name: nombre
          in: query
          description: Filtrar por nombre (like)
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageUsuarioVO'
    post:
      tags:
        - Usuarios
      summary: Crear usuario
      operationId: createUsuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioVO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioVO'
        '400':
          description: Bad Request
        '409':
          description: Conflict
  /v1/usuarios/{id}:
    get:
      tags:
        - Usuarios
      summary: Obtener usuario por ID
      operationId: getUsuarioById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioVO'
        '404':
          description: Not Found
    
    put:
      tags:
        - Usuarios
      summary: Actualizar usuario
      operationId: updateUsuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UsuarioVO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuarioVO'
        '404':
          description: Not Found
    
    delete:
      tags:
        - Usuarios
      summary: Desactivar usuario
      operationId: deleteUsuario
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content
        '404':
          description: Not Found
...
  /v1/maestros/companias:
    get:
      tags: [Maestros]
      summary: Listar compañías
      operationId: getCompanias
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompaniaVO'
    post:
      tags: [Maestros]
      summary: Crear compañía
      operationId: createCompania
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompaniaVO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompaniaVO'
  /v1/maestros/companias/{id}:
    put:
      tags: [Maestros]
      summary: Editar compañía
      operationId: updateCompania
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompaniaVO'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompaniaVO'
    delete:
      tags: [Maestros]
      summary: Eliminar compañía
      operationId: deleteCompania
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Deleted
  /v1/maestros/conceptos:
    get:
      tags: [Maestros]
      summary: Listar conceptos
      operationId: getConceptos
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ConceptoVO'
    post:
      tags: [Maestros]
      summary: Crear concepto
      operationId: createConcepto
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConceptoVO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConceptoVO'
  /v1/maestros/modelos:
    get:
      tags: [Maestros]
      summary: Listar modelos
      operationId: getModelos
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModeloVO'
    post:
      tags: [Maestros]
      summary: Crear modelo
      operationId: createModelo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModeloVO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModeloVO'
  /v1/maestros/operadoras:
    get:
      tags: [Maestros]
      summary: Listar operadoras
      operationId: getOperadoras
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OperadoraVO'
    post:
      tags: [Maestros]
      summary: Crear operadora
      operationId: createOperadora
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OperadoraVO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperadoraVO'
  /v1/maestros/parametros:
    get:
      tags: [Maestros]
      summary: Listar parametros
      operationId: getParametros
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ParametroVO'
    post:
      tags: [Maestros]
      summary: Crear parametro
      operationId: createParametro
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParametroVO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParametroVO'
  /v1/usuarios/permisos:
    get:
      tags: [Usuarios]
      summary: Listar catálogo de permisos
      operationId: getPermisosTabla
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermisoVO'
    post:
      tags: [Usuarios]
      summary: Crear permiso
      operationId: createPermiso
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PermisoVO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PermisoVO'
  /v1/usuarios/{id}/permisos:
    get:
      tags: [Usuarios]
      summary: Obtener permisos de un usuario
      operationId: getUsuarioPermisos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PermisoVO'
    post:
      tags: [Usuarios]
      summary: Asignar permisos a usuario
      operationId: updateUsuarioPermisos
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer # IDs de los permisos
                format: int64
      responses:
        '200':
          description: OK
  /v1/talleres:
    get:
      tags: [Talleres]
      summary: Listar talleres
      operationId: getTalleres
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TallerVO'
    post:
      tags: [Talleres]
      summary: Crear taller
      operationId: createTaller
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TallerVO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TallerVO'
  /v1/talleres/{id}/servicios:
    get:
      tags: [Talleres]
      summary: Listar servicios de un taller
      operationId: getTallerServicios
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ServicioTallerVO'
    post:
      tags: [Talleres]
      summary: Crear servicio en taller
      operationId: createTallerServicio
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServicioTallerVO'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServicioTallerVO'

  # --- VEHICULOS SUB-RECURSOS ---
...
components:
  schemas:
...
    UsuarioVO:
      type: object
      required:
        - nif
        - nombre
        - apellido1
        - perfilId
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        nif:
          type: string
          pattern: '^\d{8}[A-Z]$'
        nombre:
          type: string
        apellido1:
          type: string
        apellido2:
          type: string
        email:
          type: string
          format: email
        activo:
          type: boolean
          default: true
        perfilId:
          type: integer
          format: int64
        perfil:
          $ref: '#/components/schemas/PerfilVO'
        servicioAdscrito:
          type: string
    PerfilVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        activo:
          type: boolean
...
    PageUsuarioVO:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/UsuarioVO'
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
...
    CompaniaVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        contacto:
          type: string
        telefono:
          type: string
        direccion:
          type: string
        activo:
          type: boolean
    ConceptoVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        precioUnitario:
          type: number
          format: double
        kilometros:
          type: integer
        dias:
          type: integer
    ModeloVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        marcaId:
          type: integer
          format: int64
        marcaNombre:
          type: string
        tipoVehiculo:
          type: string
        activo:
          type: boolean
    OperadoraVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        telefono:
          type: string
        activo:
          type: boolean
    ParametroVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        tipoParametro:
          type: string
        activo:
          type: boolean
    PermisoVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        objeto:
          type: string
        tipoPermiso:
          type: string
          enum: [VISUALIZACION, EDICION]

    # --- NUEVOS SCHEMAS (Talleres) ---
    TallerVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        cif:
          type: string
        direccion:
          type: string
        telefono:
          type: string
        email:
          type: string
        servicios:
          type: array
          items:
            $ref: '#/components/schemas/ServicioTallerVO'
    ServicioTallerVO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        centroDirectivo:
          type: string
        servicioAdscrito:
          type: string
        nombreContacto:
          type: string
        telefonoContacto:
          type: string
        emailContacto:
          type: string
        fechaInicio:
          type: string
          format: date
        fechaFin:
          type: string
          format: date
        motivoFin:
          type: string
        numeroExpediente:
          type: string
        tiposVehiculos:
          type: array
          items:
            type: string
        conceptos:
          type: array
          items:
            type: string
        licitacion:
          type: boolean
        observaciones:
          type: string
        tipoVehiculo:
          type: string

    # --- NUEVOS SCHEMAS (Vehículos) ---
...
    LoginCertificateRequestVO:
      type: object
      properties:
        nif:
          type: string
          description: NIF recuperado del certificado
      required:
        - nif